requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
django-ninja>=1.1.0
python-jose[cryptography]
passlib[bcrypt]
email-validator>=2.0.0